GenaSQL/requirements.txt
# Core dependencies
numpy
loguru
pyyaml
python-dotenv
tenacity
func_timeout
pydantic>=2.11.3

# AI/ML dependencies
openai
cohere
boto3
google-generativeai
google-genai

# Text2SQL specific
sqlparse
sql-metadata
sqlglot
sql_metadata
spacy==3.8.3
scikit-learn
tqdm

# spacy model
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85

# extras
pandas